+++ providers/flagd/src/test/java/dev/openfeature/contrib/providers/flagd/FlagdProviderSyncResourcesCTest.java
@@ -2,9 +2,13 @@ package dev.openfeature.contrib.providers.flagd;
 
 import com.vmlens.api.AllInterleavings;
 import com.vmlens.api.Runner;
+import dev.openfeature.sdk.ErrorCode;
+import dev.openfeature.sdk.ProviderEventDetails;
 import dev.openfeature.sdk.exceptions.FatalError;
 import dev.openfeature.sdk.exceptions.GeneralError;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicReference;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -283,5 +287,95 @@ class FlagdProviderSyncResourcesCTest {
     @Test
     void firstInitializeReturnsTrue() {
         Assertions.assertTrue(flagdProviderSyncResources.initialize());
+    }
+
+    @Timeout(2)
+    @Test
+    void fatalHasPrecedenceOverInitAndShutdown() {
+        flagdProviderSyncResources.fatalError(null);
+        flagdProviderSyncResources.initialize();
+        flagdProviderSyncResources.shutdown();
+
+        Assertions.assertThrows(FatalError.class, () -> flagdProviderSyncResources.waitForInitialization(10000));
+    }
+
+    @Timeout(2)
+    @Test
+    void fatalAbortsInit() throws InterruptedException {
+        final AtomicBoolean isWaiting = new AtomicBoolean();
+        final AtomicLong waitTime = new AtomicLong(Long.MAX_VALUE);
+        final AtomicReference<Exception> fatalException = new AtomicReference<>();
+
+        Thread waitingThread = new Thread(() -> {
+            long start = System.currentTimeMillis();
+            isWaiting.set(true);
+            try {
+                flagdProviderSyncResources.waitForInitialization(10000);
+            } catch (Exception e) {
+                fatalException.set(e);
+            }
+            long end = System.currentTimeMillis();
+            long duration = end - start;
+            waitTime.set(duration);
+        });
+        waitingThread.start();
+
+        while (!isWaiting.get()) {
+            Thread.yield();
+        }
+
+        Thread.sleep(MAX_TIME_TOLERANCE); // waitingThread should have started waiting in the meantime
+
+        var fatalEvent = ProviderEventDetails.builder()
+                .errorCode(ErrorCode.PROVIDER_FATAL)
+                .message("Some message")
+                .build();
+        flagdProviderSyncResources.fatalError(fatalEvent);
+
+        waitingThread.join();
+
+        var wait = MAX_TIME_TOLERANCE * 3;
+
+        Assertions.assertTrue(
+                waitTime.get() < wait,
+                () -> "Wakeup should be almost instant, but took " + waitTime.get()
+                        + " ms, which is more than the max of"
+                        + wait + " ms");
+        Assertions.assertNotNull(fatalException.get());
+        Assertions.assertInstanceOf(FatalError.class, fatalException.get());
+        Assertions.assertEquals(
+                "Initialization failed due to a fatal error: " + fatalEvent.getMessage(),
+                fatalException.get().getMessage());
+    }
+
+    @Timeout(2)
+    @Test
+    void callingShutdownWithPreviousNonFatal_wakesUpWaitingThread_WithGeneralException() throws InterruptedException {
+        final AtomicBoolean isWaiting = new AtomicBoolean();
+        final AtomicBoolean successfulTest = new AtomicBoolean();
+
+        Thread waitingThread = new Thread(() -> {
+            long start = System.currentTimeMillis();
+            isWaiting.set(true);
+            Assertions.assertThrows(GeneralError.class, () -> flagdProviderSyncResources.waitForInitialization(10000));
+
+            long end = System.currentTimeMillis();
+            long duration = end - start;
+            var wait = MAX_TIME_TOLERANCE * 3;
+            successfulTest.set(duration < wait);
+        });
+        waitingThread.start();
+
+        while (!isWaiting.get()) {
+            Thread.yield();
+        }
+
+        Thread.sleep(MAX_TIME_TOLERANCE); // waitingThread should have started waiting in the meantime
+
+        flagdProviderSyncResources.shutdown();
+
+        waitingThread.join();
+
+        Assertions.assertTrue(successfulTest.get());
     }
 }
